LayoutTests/imported/w3c/web-platform-tests/custom-elements/scoped-registry/ShadowRoot-init-registry.tentative-expected.txt


FAIL ShadowRoot.registry is null if not explicitly specified assert_equals: expected (object) null but got (undefined) undefined
FAIL Attach the global registry to a shadow root assert_equals: expected (object) object "[object CustomElementRegistry]" but got (undefined) undefined
FAIL Attach a non-global registry to a shadow root assert_equals: expected (object) object "[object CustomElementRegistry]" but got (undefined) undefined
FAIL Attach the same registry to multiple shadow roots assert_equals: expected (object) object "[object CustomElementRegistry]" but got (undefined) undefined
FAIL Attaching registry to shadow root can only be done during initialization assert_equals: expected null but got object "[object CustomElementRegistry]"
